Move prefix utilities to utils/rdf/ directory
