+++ packages/graph-explorer/src/utils/rdf/common-prefixes.json
