+++ packages/graph-explorer/src/utils/rdf/generatePrefixes.test.ts
