+++ packages/graph-explorer/src/utils/rdf/generatePrefixes.ts
