+++ packages/graph-explorer/src/utils/rdf/index.ts
@@ -1,0 +1,2 @@
+export { default as replacePrefixes } from "./replacePrefixes";
+export { default as generatePrefixes } from "./generatePrefixes";
