+++ packages/graph-explorer/src/utils/rdf/replacePrefixes.test.ts
