+++ packages/graph-explorer/src/utils/rdf/replacePrefixes.ts
